+++ herzogdupont.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * @package   Herzog Dupont
+ * @package   Herzog Dupont for YOOtheme Pro
  * @author    Thomas Weidlich
  * @copyright Copyright (C) 2021 Thomas Weidlich
  * @license   GNU General Public License version 3, see LICENSE