+++ modules/elements/elements/hd-flipcard/templates/content.php
@@ -1,6 +1,6 @@
 <?php
 
-/* Herzog Dupont Copyright (C) 2016–2021 YOOtheme GmbH, 2019–2021 Thomas Weidlich GNU GPL v3 */
+/* Herzog Dupont for YOOtheme Pro Copyright (C) 2016–2021 YOOtheme GmbH, 2019–2021 Thomas Weidlich GNU GPL v3 */
 
 if ($props['image'] || $props['title'] || $props['meta'] || $props['content'] || $props['link']) : ?>
 <div>