+++ modules/elements/elements/hd-flipcard/templates/template-content_back.php
@@ -1,13 +1,13 @@
 <?php
 
-/* Herzog Dupont Copyright (C) 2016–2021 YOOtheme GmbH, 2019–2021 Thomas Weidlich GNU GPL v3 */
+/* Herzog Dupont for YOOtheme Pro Copyright (C) 2016–2021 YOOtheme GmbH, 2019–2021 Thomas Weidlich GNU GPL v3 */
 
 // Title
 $title_back = $this->el($props['title_back_element'], [
 
     'class' => [
         'el-title-back',
-        'uk-{title_back_style}',
+        'uk-[text-{@title_back_style: meta|lead}]{title_back_style}',
         'uk-card-title {@panel_back_style} {@!title_back_style}',
         'uk-heading-{title_back_decoration}',
         'uk-font-{title_back_font_family}',
@@ -24,10 +24,10 @@ $meta_back = $this->el($props['meta_back_element'], [
 
     'class' => [
         'el-meta-back',
-        'uk-[text-{@meta_back_style: meta}]{meta_back_style}',
+        'uk-[text-{@meta_back_style: meta|lead}]{meta_back_style}',
         'uk-text-{meta_back_color}',
         'uk-margin[-{meta_back_margin}]-top {@!meta_back_margin: remove}',
-        'uk-margin-remove-bottom [uk-margin-{meta_back_margin: remove}-top]' => !in_array($props['meta_back_style'], ['', 'meta']) || $props['meta_back_element'] != 'div',
+        'uk-margin-remove-bottom [uk-margin-{meta_back_margin: remove}-top]' => !in_array($props['meta_back_style'], ['', 'meta', 'lead']) || $props['meta_back_element'] != 'div',
     ],
 
 ]);
@@ -37,12 +37,13 @@ $content_back = $this->el('div', [
 
     'class' => [
         'el-content-back uk-panel',
-        'uk-text-{content_back_style}',
+        'uk-[text-{@content_back_style: meta|lead}]{content_back_style}',
         '[uk-text-left{@content_back_align}]',
         'uk-dropcap {@content_back_dropcap}',
         'uk-column-{content_back_column}[@{content_back_column_breakpoint}]',
         'uk-column-divider {@content_back_column} {@content_back_column_divider}',
         'uk-margin[-{content_back_margin}]-top {@!content_back_margin: remove}',
+        'uk-margin-remove-bottom [uk-margin-{content_back_margin: remove}-top]' => !in_array($props['content_back_style'], ['', 'meta', 'lead']),
     ],
 
 ]);